remove cli_test_suite, update canary timeouts
