+++ yarn-project/aztec-cli/package.json
@@ -45,7 +45,6 @@
     "commander": "^9.0.0",
     "lodash.startcase": "^4.4.0",
     "semver": "^7.5.4",
-    "string-argv": "^0.3.2",
     "tslib": "^2.4.0",
     "viem": "^1.2.5"
   },
