+++ yarn-project/aztec-cli/src/index.ts
@@ -34,8 +34,6 @@ import {
   prepTx,
 } from './utils.js';
 
-export { cliTestSuite } from './test/cli_test_suite.js';
-
 const accountCreationSalt = Fr.ZERO;
 
 const stripLeadingHex = (hex: string) => {
