--- yarn-project/canary/src/web/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <body>
-    <script type="module">
-      import * as AztecJs from '@aztec.aztec.js';
-
-      window.AztecJs = AztecJs;
-    </script>
-  </body>
-</html>
